Open an XlsxFile from a slice of bytes
New interface added to be able to read an
XlsxFile from a slice of bytes. Useful for
when the file is already in memory and not
reading of disk.

Upated the naming of the *File struct replaced
Excel with Xlsx in attempt to this generic Xlsx reader
and not tied to any product/platform

Ref TDR-493