Support Raw Date Type
Raw date types appear in the XML as a cell, with type "d".
Previously we assumed if the style indicated it was a date,
that the type was a number, and an error occurred trying to
parse the cell value as a number.

Instead, we now check if the value is a number as well as if's a date
format.
If it is of type "d", we therefore simply return the raw value we are given.