Parse all string values from shared strings
Some values appear in the shared strings file as
<si> <t> value </t> </si>
others appear as
<si> <r> <t> value </t> </r> </si>

We were only picking up the first set of values, which
led to the second set being completely ignored. This would lead
to elements beyond the end of the shared strings being attempted
to be read, and incorrect index values for a lot of the elements,
meaning the data presented was not correct.

The shared strings reader has been expanded to parse both cases,
and then populate, as per the existing structure, a single array of
strings.