Parse multiple RichText values from shared strings
Some of the shared strings can contain multiple <r><t> elements
e.g.
<si> <r> <t> some </t> </r> <r> <t> value </r> </t> </si>

In order to correctly determine the entire contents of the string
it is necessary to parse all these values, and then combine them
into a single value.

The previous iteration of the sharedStrings parser did not handle
multiple r and t elements, so it has been updated to correctly
build the content in these cases.