Update Cell parser to handle shared strings with a date format style
In addition to all the other funky date things,
it's possible to have a date cell, but with the text value stored
as a sharedString. In this case the type will be 's' but the
style will indicate that it is a date.

In this case, we want to read the value from sharedStrings,
rather than trying to parse it as an actual date.