Expose the xlsx sheetFile map
Rather than simply export the map, I've added a function that allows the user to supply the already
exported sheetname, and get the corresponding zip.File for it.

This is particularly useful when the user wants to read some data from the underlying XML that is not read by the library.